backend/requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
requests==2.32.3
pydantic==2.11.4
pydantic-settings==2.9.1
asyncpg==0.30.0
fastapi==0.116.1
pydantic_core===2.33.2
PyJWT==2.10.1
SQLAlchemy==2.0.42
starlette==0.47.2
pydantic[email]>=2.0
mangum==0.17.0
loguru==0.7.3
gspread==5.12.0

pydantic_core~=2.33.2
idna~=3.11
anyio~=4.11.0
sniffio~=1.3.1
typing_extensions~=4.15.0
annotated-types~=0.7.0
attrs~=25.4.0
h11~=0.16.0
pip~=24.3.1
click~=8.3.0
uvicorn~=0.38.0
jose~=1.0.0
passlib~=1.7.4
pydantic_core~=2.33.2
dnspython~=2.8.0
cryptography~=46.0.3
rsa~=4.9.1
pyasn1~=0.6.1
cffi~=2.0.0
python-jose~=3.5.0
ecdsa~=0.19.1
six~=1.17.0
typing_extensions~=4.15.0
email-validator~=2.3.0
greenlet~=3.2.4
pycparser~=2.23
bcrypt~=4.1.2
python-multipart~=0.0.20
pytz~=2025.2
Authlib~=1.6.5